Add resizeobserver check to all overview page charts
